Update: allows to train several architectures at the same time
-Launch the training with the config.py and lib/model.py files, modifying architecture, loss, inputs etc.
-When starting from scratch, we write the correspondent conf and mconf dicts into the modelDir folder.
Those are loaded when restarting, allowing to keep the wanted config in a different file from the default config.py.
-We do the same for the architecture, copying the file model.py into the same modelDir directory.
When restarting, we copy temporarily this file into /lib and import it as a module, to train on the same architecture as
the one that is saved in the model_dict.
-We do the same for the print_output.py utility, that loads the weights and allows to visualize output, targets and
erros.